V13/gestor_team/requirements.txt
aiohttp>=3.5.1
appdirs>=1.4.3
argh>=0.26.2
async-timeout>=3.0.1
attrs>=18.2.0
Babel>=2.6.0
beautifulsoup4>=4.7.1
cached-property>=1.5.1
certifi>=2018.8.24
chardet>=3.0.4
decorator>=4.3.0
defusedxml>=0.5.0
docutils>=0.14
feedparser>=5.2.1
freezegun>=0.3.11
gevent>=1.3.7
greenlet>=0.4.15
html2text>=2018.1.9
idna>=2.6
isodate>=0.6.0
Jinja2>=2.10
libsass>=0.17.0
lxml>=4.3.2
Mako>=1.0.7
MarkupSafe>=1.1.0
matplotlib>=3.3.3
mock>=2.0.0
mpld3>=0.5.2
multidict>=4.5.2
num2words>=0.5.6
numpy>=1.19.4
odoo>=13.0.post20200629
ofxparse>=0.19
openpyxl>=3.0.5
pandas>=1.1.4
passlib>=1.7.1
pathtools>=0.1.2
pbr>=4.2.0
phonenumbers>=8.9.10
Pillow>=5.4.1
polib>=1.1.0
psutil>=5.5.1
psycopg2>=2.7.7
pyasn1>=0.4.2
pyasn1-modules>=0.2.1
pydot>=1.4.1
pymssql>=2.1.5
pyodbc>=4.0.30
pyparsing>=2.2.0
PyPDF2>=1.26.0
pyserial>=3.4
python-dateutil>=2.7.3
python-ldap>=3.1.0
python-slugify>=2.0.1
python-stdnum>=1.10
pytz>=2019.1
pyusb>=1.0.2
pyxlsb>=1.0.6
PyYAML>=3.13
qrcode>=6.1
reportlab>=3.5.13
requests>=2.21.0
requests-toolbelt>=0.8.0
roman>=2.0.0
six>=1.12.0
soupsieve>=1.8
suds-jurko
sqlalchemy
Unidecode>=1.0.23
urllib3>=1.24.1
vatnumber>=1.2
vobject>=0.9.6.1
watchdog>=0.9.0
wheel>=0.36.2
Werkzeug>=0.14.1
xlrd>=2.0.1
XlsxWriter>=1.1.2
xlwt>=1.3.0
yarl>=1.3.0
zeep>=3.2.0
# Desde rama DEV publicada
# Cambios en DEV con cambio en action
# con la corrección en la action
# mas cambios en Dev
# Otro mas con este comando: git branch --set-upstream-to=origin/Dev
# Más cambios
# otro cambio
# corrigiendo
# mas pruebas desde VSC
#comentario
# prueba de operaciób en dev